requirements.txt
fastapi
uvicorn
openai
pika
boto3
python-dotenv
requests